+++ Tests/Realm.Tests/Sync/UserManagementTests.cs
@@ -148,6 +148,35 @@ namespace Realms.Tests.Sync
         }
 
         [Test]
+        public void AppDeleteUserFromServer_RemovesUser()
+        {
+            SyncTestHelpers.RunBaasTestAsync(async () =>
+            {
+                var username = SyncTestHelpers.GetVerifiedUsername();
+                var password = SyncTestHelpers.DefaultPassword;
+                var user = await GetUserAsync(app: null, username, password);
+
+                Assert.That(DefaultApp.CurrentUser, Is.EqualTo(user));
+
+                await DefaultApp.DeleteUserFromServerAsync(user);
+                Assert.That(DefaultApp.CurrentUser, Is.Null);
+
+                Exception ex = null;
+                try
+                {
+                    await DefaultApp.LogInAsync(Credentials.EmailPassword(username, password));
+                }
+                catch (Exception e)
+                {
+                    ex = e;
+                }
+
+                Assert.That(ex, Is.TypeOf<AppException>());
+                Assert.That(ex.Message, Is.EqualTo("InvalidPassword: invalid username/password"));
+            });
+        }
+
+        [Test]
         public void EmailPasswordRegisterUser_Works()
         {
             SyncTestHelpers.RunBaasTestAsync(async () =>